fixing entities and halfway done product crud

+++ src/main/java/com/example/ecommerce_backend/service/implementations/UserServiceImpl.java
@@ -2,8 +2,13 @@ package com.example.ecommerce_backend.service.implementations;
 
 import com.example.ecommerce_backend.dto.UserEntity.UserEntityCreateDto;
 import com.example.ecommerce_backend.dto.UserEntity.UserEntityIndexDto;
+import com.example.ecommerce_backend.dto.UserEntity.UserEntityUpdateDto;
+import com.example.ecommerce_backend.exception.ResourceDuplicateException;
+import com.example.ecommerce_backend.exception.ResourceNotFoundException;
 import com.example.ecommerce_backend.mapper.UserEntity.UserEntityCreateDtoMapper;
 import com.example.ecommerce_backend.mapper.UserEntity.UserEntityIndexDtoMapper;
+import com.example.ecommerce_backend.mapper.UserEntity.UserEntityUpdateDtoMapper;
+import com.example.ecommerce_backend.model.SupplierEntity;
 import com.example.ecommerce_backend.model.UserEntity;
 import com.example.ecommerce_backend.repository.UserEntityRepository;
 import com.example.ecommerce_backend.service.interfaces.UserServiceInterface;
@@ -26,32 +31,51 @@ public class UserServiceImpl implements UserServiceInterface {
     private final UserEntityRepository userEntityRepository;
     private final UserEntityCreateDtoMapper userEntityCreateDtoMapper;
     private final UserEntityIndexDtoMapper userEntityIndexDtoMapper;
+    private final UserEntityUpdateDtoMapper userEntityUpdateDtoMapper;
     private final PasswordEncoder passwordEncoder;
 
     @Override
-    public Optional<UserEntity> registerUser(UserEntityCreateDto userDto) {
-        UserEntity userEntity = userEntityCreateDtoMapper.UserEntityCreateDtoToUserEntity(userDto);
-        userEntity.setPassword(passwordEncoder.encode(userEntity.getPassword()));
-        userEntity = userEntityRepository.save(userEntity);
-        return Optional.of(userEntity);
+    public UserEntity registerUser(UserEntityCreateDto userDto) {
+        if (!userEntityRepository.existsByUsername(userDto.getUsername())) {
+            UserEntity userEntity = userEntityCreateDtoMapper.UserEntityCreateDtoToUserEntity(userDto);
+            userEntity.setPassword(passwordEncoder.encode(userEntity.getPassword()));
+            userEntity = userEntityRepository.save(userEntity);
+            return userEntity;
+        } else {
+            throw new ResourceDuplicateException("username already exist");
+        }
     }
 
     public void deleteUserById(int id) {
-        userEntityRepository.deleteById(id);
+        if (userEntityRepository.existsById(id)) {
+            userEntityRepository.deleteById(id);
+        } else {
+            throw new ResourceNotFoundException("Could not find user with id " + id);
+        }
     }
 
-    public Optional<UserEntity> findUserById(int id) {
-        return userEntityRepository.findById(id);
+    public UserEntity findUserById(int id) {
+        Optional<UserEntity> userEntity = userEntityRepository.findById(id);
+        if (userEntity.isPresent()) {
+            return userEntity.get();
+        } else {
+            throw new ResourceNotFoundException("Could not find user with id " + id);
+        }
     }
 
     public Page<UserEntityIndexDto> findByCondition(Pageable pageable) {
         return userEntityRepository.findAll(pageable).map(userEntityIndexDtoMapper::UserEntityToUserEntityIndexDto);
     }
 
-//    public UserEntityDto update(UserEntityDto userDto, int id) {
-//        UserEntityDto data = findById(id);
-//        User entity = userMapper.toEntity(userDto);
-//        BeanUtil.copyProperties(data, entity);
-//        return save(userMapper.toDto(entity));
-//    }
+    public UserEntity updateUser(UserEntityUpdateDto userEntityUpdateDto) {
+        Optional<UserEntity> userEntity = userEntityRepository.findById(userEntityUpdateDto.getId());
+        if (userEntity.isPresent()) {
+            userEntity.get().setFirstName(userEntityUpdateDto.getFirstName());
+            userEntity.get().setLastName(userEntityUpdateDto.getLastName());
+            userEntity.get().setTelephone(userEntityUpdateDto.getTelephone());
+            return userEntityRepository.save(userEntity.get());
+        } else {
+            throw new ResourceNotFoundException("Could not find user with id " + userEntityUpdateDto.getId());
+        }
+    }
 }
